+++ weblate/screenshots/forms.py
@@ -17,19 +17,44 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 
-
 from django import forms
 
 from weblate.screenshots.models import Screenshot
 from weblate.trans.forms import QueryField
+from weblate.utils.forms import SortedSelect
 
 
-class ScreenshotForm(forms.ModelForm):
-    """User profile editing."""
+class ScreenshotEditForm(forms.ModelForm):
+    """Screenshot editing."""
 
     class Meta:
         model = Screenshot
         fields = ("name", "image")
+
+
+class LanguageChoiceField(forms.ModelChoiceField):
+    def label_from_instance(self, obj):
+        return obj.language
+
+
+class ScreenshotForm(forms.ModelForm):
+    """Screenshot upload."""
+
+    class Meta:
+        model = Screenshot
+        fields = ("name", "image", "translation")
+        widgets = {
+            "translation": SortedSelect,
+        }
+        field_classes = {
+            "translation": LanguageChoiceField,
+        }
+
+    def __init__(self, component, data=None, files=None, instance=None):
+        self.component = component
+        super().__init__(data=data, files=files, instance=instance)
+        self.fields["translation"].queryset = component.translation_set.all()
+        self.fields["translation"].initial = component.source_translation
 
 
 class SearchForm(forms.Form):