+++ weblate/screenshots/views.py
@@ -28,7 +28,7 @@ from django.views.decorators.http import require_POST
 from django.views.generic import DetailView, ListView
 from PIL import Image
 
-from weblate.screenshots.forms import ScreenshotForm, SearchForm
+from weblate.screenshots.forms import ScreenshotEditForm, ScreenshotForm, SearchForm
 from weblate.screenshots.models import Screenshot
 from weblate.trans.models import Unit
 from weblate.utils import messages
@@ -49,9 +49,7 @@ def try_add_source(request, obj):
         return False
 
     try:
-        source = obj.component.source_translation.unit_set.get(
-            pk=int(request.POST["source"])
-        )
+        source = obj.translation.unit_set.get(pk=int(request.POST["source"]))
     except (Unit.DoesNotExist, ValueError):
         return False
 
@@ -66,7 +64,9 @@ class ScreenshotList(ListView, ComponentViewMixin):
 
     def get_queryset(self):
         self.kwargs["component"] = self.get_component()
-        return Screenshot.objects.filter(component=self.kwargs["component"]).order()
+        return Screenshot.objects.filter(
+            translation__component=self.kwargs["component"]
+        ).order()
 
     def get_context_data(self, **kwargs):
         result = super().get_context_data(**kwargs)
@@ -76,17 +76,17 @@ class ScreenshotList(ListView, ComponentViewMixin):
             if self._add_form is not None:
                 result["add_form"] = self._add_form
             else:
-                result["add_form"] = ScreenshotForm()
+                result["add_form"] = ScreenshotForm(component)
         return result
 
     def post(self, request, **kwargs):
         component = self.get_component()
         if not request.user.has_perm("screenshot.add", component):
             raise PermissionDenied()
-        self._add_form = ScreenshotForm(request.POST, request.FILES)
+        self._add_form = ScreenshotForm(component, request.POST, request.FILES)
         if self._add_form.is_valid():
             obj = Screenshot.objects.create(
-                component=component, user=request.user, **self._add_form.cleaned_data
+                user=request.user, **self._add_form.cleaned_data
             )
             request.user.profile.increase_count("uploaded")
 
@@ -111,23 +111,25 @@ class ScreenshotDetail(DetailView):
 
     def get_object(self, *args, **kwargs):
         obj = super().get_object(*args, **kwargs)
-        self.request.user.check_access_component(obj.component)
+        self.request.user.check_access_component(obj.translation.component)
         return obj
 
     def get_context_data(self, **kwargs):
         result = super().get_context_data(**kwargs)
-        component = result["object"].component
+        component = result["object"].translation.component
         if self.request.user.has_perm("screenshot.edit", component):
             if self._edit_form is not None:
                 result["edit_form"] = self._edit_form
             else:
-                result["edit_form"] = ScreenshotForm(instance=result["object"])
+                result["edit_form"] = ScreenshotEditForm(instance=result["object"])
         return result
 
     def post(self, request, **kwargs):
         obj = self.get_object()
-        if request.user.has_perm("screenshot.edit", obj.component):
-            self._edit_form = ScreenshotForm(request.POST, request.FILES, instance=obj)
+        if request.user.has_perm("screenshot.edit", obj.translation):
+            self._edit_form = ScreenshotEditForm(
+                request.POST, request.FILES, instance=obj
+            )
             if self._edit_form.is_valid():
                 if request.FILES:
                     obj.user = request.user
@@ -142,10 +144,11 @@ class ScreenshotDetail(DetailView):
 @login_required
 def delete_screenshot(request, pk):
     obj = get_object_or_404(Screenshot, pk=pk)
-    if not request.user.has_perm("screenshot.delete", obj.component):
+    component = obj.translation.component
+    if not request.user.has_perm("screenshot.delete", obj.translation):
         raise PermissionDenied()
 
-    kwargs = {"project": obj.component.project.slug, "component": obj.component.slug}
+    kwargs = {"project": component.project.slug, "component": component.slug}
 
     obj.delete()
 
@@ -156,7 +159,7 @@ def delete_screenshot(request, pk):
 
 def get_screenshot(request, pk):
     obj = get_object_or_404(Screenshot, pk=pk)
-    if not request.user.has_perm("screenshot.edit", obj.component):
+    if not request.user.has_perm("screenshot.edit", obj.translation.component):
         raise PermissionDenied()
     return obj
 
@@ -197,7 +200,7 @@ def search_results(code, obj, units=None):
 @require_POST
 def search_source(request, pk):
     obj = get_screenshot(request, pk)
-    translation = obj.component.source_translation
+    translation = obj.translation
 
     form = SearchForm(request.POST)
     if not form.is_valid():
@@ -225,7 +228,7 @@ def ocr_search(request, pk):
     obj = get_screenshot(request, pk)
     if not HAS_OCR:
         return search_results(500, obj)
-    translation = obj.component.source_translation
+    translation = obj.translation
 
     # Load image
     original_image = Image.open(obj.image.path)